Corrected some mistakes

+++ backend/app/models/user.py
@@ -5,5 +5,5 @@ class User(db.Model):
     prenom = db.Column(db.String(25), nullable=False)
     nom = db.Column(db.String(25), nullable=False)
     password = db.Column(db.String(100), nullable=False)
-    numeroTel = db.Column(db.Integer(15), nullable=False)
+    numeroTel = db.Column(db.Integer, nullable=False)
     dateNaissance = db.Column(db.Date, nullable=False)
